requirements.txt
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.1.0
langchain-experimental>=0.1.0
openai>=1.0.0